packages/@aws-cdk/aws-lambda-python/test/lambda-handler-custom-build-docker-image/requirements.txt
# Lock versions of pip packages
certifi==2020.6.20
chardet==3.0.4
idna==2.10
urllib3==1.25.11
# Requests used by this lambda
requests==2.23.0
# Pillow 6.x so that python 2.7 and 3.x can both use this fixture
pillow==6.2.2
